cleaned up unused packages, Makefiles
