+++ sites/zappa/app/controllers.py
@@ -1,13 +1,15 @@
 import bisect
 import copy
 import urllib
+import datetime
 import sys
 import os
 import json
+import os.path
 import requests
-from datetime import datetime
 
-from .errors import GitlabError, EnvironmentVariableError, MalformedFusilladeConfigError
+import hca
+from dcplib.aws.clients import secretsmanager as sm_client  # type: ignore
 
 
 class FileController:
@@ -24,18 +26,34 @@ class FileController:
             try:
                 assert group in modified_data.get("groups")
             except AssertionError:
-                raise MalformedFusilladeConfigError(f'group "{group}" not found in {modified_data.get("groups")} ')
-            group_data = modified_data['groups'][group]
-            users_in_current_group = group_data.get('users',[])
+                print(f'group "{group}" not found in {modified_data.get("groups")} ')
+                continue
+            users_in_current_group = modified_data['groups'][group]['users']
             if user not in users_in_current_group:
                 bisect.insort(users_in_current_group, user)
         return modified_data
+
+class FusilladeConfig:
+    def __init__(self):
+        self.api_endpoint = os.getenv('FUS_API_ENDPOINT')
+        self.swagger_endpoint = f'{self.api_endpoint}/swagger.json'
+        self.auth_client = hca.auth.AuthClient(swagger_url=self.swagger_endpoint)
+
+    def get_users(self):
+        return self.auth_client.get_v1_users()
+
+    def get_groups(self):
+        return self.auth_client.get_v1_groups()
+
+    def get_role(self):
+        return self.auth_client.get_v1_roles()
+
 
 class GitlabController:
 
     def __init__(self):
         self.access_token = os.getenv('DCP_FUS_GL_TOKEN')
-        self.gitlab_api_url  = os.getenv("GITLAB_API_URL")
+        self.gitlab_url  = os.getenv("GITLAB_URL")
         self.access_headers = {"PRIVATE-TOKEN": self.access_token}
         self.gitlab_project_id = os.getenv("GITLAB_PROJECT_ID")
         self.ci_branch = self._get_ci_branch()
@@ -45,63 +63,80 @@ class GitlabController:
     def get_file_from_repo(self, file_path: os.PathLike) -> requests.Response:
         """ get a file from a repo, file_path is from root of repository"""
         encoded_file_path = urllib.parse.quote(file_path, safe='')
-        gl_file_path = f'/projects/{self.gitlab_project_id}/repository/files/{encoded_file_path}/raw'
-        url = f'{self.gitlab_api_url}{gl_file_path}'
+        gl_file_path = f'/api/v4/projects/{self.gitlab_project_id}/repository/files/{encoded_file_path}/raw'
+        url = f'{self.gitlab_url}{gl_file_path}'
         parameters = {"ref": self.ci_branch}
         resp = requests.get(url=url, headers=self.access_headers, params=parameters)
         return resp
+
+    def create_branch(self,service_account_name):
+        gl_branch_path = f'/projects/{self.gitlab_project_id}/repository/branches'
+        parameters = {"branch": self._new_branch_name,
+                      "ref": self.ci_branch}
+        url = f'{self.gitlab_url}{gl_branch_path}'
+        try:
+            r = requests.post(url=url,headers=self.access_headers,params=parameters)
+            r.raise_for_status()
+        except requests.exceptions.HTTPError as err:
+            print(err)
+            exit(1)
+        return r.json()
+
+    def get_gitlab_access_token(self):
+        secret_store = os.getenv('FUS_SECRETS_STORE')
+        secret_name = os.getenv('GITLAB_ACCESS_KEY_SECRET_NAME')
+        return sm_client.get_secret_value(SecretId=f'{secret_store}/{secret_name}').get('SecretString')
+
 
     # https://docs.gitlab.com/ee/api/commits.html
     def commit_changes_(self, service_account_name, modified_file: FileController):
         gl_commit_path = f'/projects/{self.gitlab_project_id}/repository/commits'
         commit_message = f'This commit was created from DCP-Fusillade'
-        actions = [{'action': 'update',
+        actions = [{'action':'update',
                     'file_path': modified_file.file_path,
-                    'content': json.dumps(modified_file.updated_data, indent=4)}]
+                    'content': modified_file.updated_data}]
         # Note this might be able to be reduced to 1 api call to gitlab.
         # may need to set force to true.
         payload = {'branch': self._new_branch_name,
-                   'start_branch': self.ci_branch,
                    'commit_message': commit_message,
-                   'force': False,
-                   'actions': actions}
-        url = f'{self.gitlab_api_url}{gl_commit_path}'
+                   'force': True,
+                   'actions':actions}
+        url = f'{self.gitlab_url}{gl_commit_path}'
         headers = self.access_headers
         headers['Content-Type']= 'application/json'
         try:
             r = requests.post(url, headers=headers, data=json.dumps(payload))
             r.raise_for_status()
         except requests.exceptions.HTTPError as err:
-            msg = "Error calling Gitlab URL to commit propposed changes "
-            msg += f"to branch ({self._new_branch_name}): "
-            msg += str(err)
-            raise GitlabError(msg)
+            print(err)
+            exit(1)
         return r.json()
 
-    def create_merge_request(self):
+    def create_merge_path(self):
         gl_merge_path = f"/projects/{self.gitlab_project_id}/merge_requests"
         payload = {"id": self.gitlab_project_id,
                    "source_branch": self._new_branch_name,
                    "target_branch": self.ci_branch,
                    "title": f"Request to add user to groups",
-                   "description": "This pull request was opened automatically.",
                    "remove_source_branch": True,
-                   "allow_collaboration": True,
                    "squash": True}
-        url = f'{self.gitlab_api_url}{gl_merge_path}'
+        url = f'{self.gitlab_url}{gl_merge_path}'
         headers = self.access_headers
         headers['Content-Type'] = 'application/json'
         try:
             r = requests.post(url, headers=headers, data=json.dumps(payload))
             r.raise_for_status()
         except requests.exceptions.HTTPError as err:
-            msg = f"Error calling Gitlab URL to commit changes: {url} "
-            msg += "Error Message: "
-            msg += str(err)
-            raise GitlabError(msg)
+            print(err)
+            exit(1)
         return r.json()
+
+
 
     @staticmethod
     def _get_ci_branch():
         stage = os.getenv('FUS_DEPLOYMENT_STAGE')
         return stage if stage != 'dev' else 'master'
+
+
+
